handle marriages fact dates and place

+++ src/main/java/com/rootscity/model/hints/Hint.java
@@ -7,6 +7,6 @@ public class Hint {
   public String personaId;        // id of the persona (not the whole URI)
   public String recordId;         // id of the record that the persona belongs to
   public String collectionTitle;  // title of the collection that the record belongs to
-  public double confidence;      // 1-5, 5 being "most confident" (by default you will see only values from 3-5)
+  public Integer confidence;      // 1-5, 5 being "most confident" (by default you will see only values from 3-5)
   public Integer eventYear;
 }
